realTimeScanning code organized from main.ts
