+++ src/interfaces/KanbanView.ts
@@ -50,8 +50,8 @@ export const DEFAULT_SETTINGS: GlobalSettings = {
 			ignoreFileNameDates: false,
 			taskCompletionFormat: "ObsidianTasks",
 			taskCompletionInLocalTime: true,
-			taskCompletionShowUtcOffset: true,
-			autoAddDue: true,
+			taskCompletionShowUtcOffset: false,
+			autoAddDue: false,
 			scanVaultAtStartup: false,
 			dayPlannerPlugin: false,
 			realTimeScanning: false,
